Fixes for CI release workflow (#8)
* Fixes for CI release workflow

The release workflow has been updated to prefix the generated version tag with "v", as well as to allow manual invocation.

* Updated PR workflow criteria

I updated the PR workflow to run at all times rather than only when certain files are modified. This is because this is listed as a mandatory check for pull requests, so it has to run.