Fixed message for failed subnet balance test
The failure message for the test asserting that EC2 instances are balanced in a set of subnets has been fixed so that it will display the actual instance ID as well as the actual subnet IDs (both actual and expected). Previously the message did not include the instance ID, and the actual and expected values were output as pointers rather than values.