chore: upgrade go and various deps
This updates the required Go version from 1.17 to 1.21, as well as
updating numerous dependencies to their newest version.

- github.com/aws/aws-sdk-go-v2/service/dax v1.6.0 -> v1.17.6
- github.com/aws/aws-sdk-go-v2/service/ec2 v1.142.1 -> v1.146.0
- github.com/aws/aws-sdk-go-v2/service/eks v1.18.0 -> v1.37.1
- github.com/aws/aws-sdk-go-v2/service/iam v1.11.0 -> v1.28.7
- github.com/aws/aws-sdk-go-v2/service/route53 v1.36.1 -> v1.37.1
- github.com/golangci/golangci-lint v1.46.2 -> v1.56.0
- github.com/gruntwork-io/terratest v0.38.8 -> v0.46.11
- github.com/stretchr/testify v1.7.1 -> v1.8.4
- golang.org/x/tools v0.1.11-0.20220316014157-77aa08bb151a -> v0.17.0
- k8s.io/api v0.23.1 -> v0.29.1
- k8s.io/apimachinery v0.23.1 -> v0.29.1
- k8s.io/client-go v0.23.1 -> v0.29.1
- sigs.k8s.io/aws-iam-authenticator v0.5.3 -> v0.6.17

The linter configuration has been fixed to remove several
deprecated linters (varcheck, deadcode, and structcheck).

The version of various GitHub Actions used in the CI process has been
updated to newer versions, as well as removing the no-longer-required
cache action (the newer setup-go action takes care of this).